remove getIntialProps from events
